platform/reusableIntelligentServices/kiServices.functions/src/main/python/qrScan/requirements.txt
pyzbar
opencv-python
numpy
Pillow